Update patch set 4
Patch Set 4:

One problem I see here is that if later you need to add another piece to the mirror process (e.g. another file to download from some URL) you'll have to manually go in and delete the `done` file for each of the versions.
Another option would be to make the `done` file a list of items downloaded, and if the item is *not* found in the done file, download it independently, in other words, make the check more granular.

Patch-set: 4