Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

we also need to copy the data recursively as the command is creating a hiearchy of files and directories. and then this directory needs to be stored a tarball and removed for the upload process to work.

Patch-set: 2
Reviewer: Gerrit User 20 <20@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1