Update patch set 1
Patch Set 1:

You're right, it's not longer necessary to create a custom ssh config. The inventory now has the reference to the ssh key.
Do you think it may be worth leaving a custom ssh config for the cluster, just in case someone wants to connect to it directly from the jumphost ala `ssh -F ~/.ssh/cluster-config master-N`?

Patch-set: 1