Update patch set 2
Patch Set 2:

Indeed, I found the issue when rebuilding my libvirt lab (connected), :P
On my previous comment I was focusing in the `when` conditional, as it seems duplicated because once it enters in the role it will evaluate again if the provision_cache_store is defined. But I was also mentioning that this way is a quicker way to bypass it.

Patch-set: 2