Update patch set 13
Patch Set 13: Code-Review+2

Changes look good to me, the tests is the only part that was skipped in Dallas, but I see the pipeline for 4.4 has commented the `cnf_tests_mode`.
I think it may be worth running the check with 4.5 where it is set to offline so the tests are validated.

Patch-set: 13
Label: Code-Review=+2