Add a generic role for deploying operators
When deploying and operator based product, the first step is to ask
OLM to run its operator. This step is the same for all the products,
so it makes sense to have a generic role for it.

The role can be used for different operators by calling it with
different parameters.

Change-Id: Id073105badd43062ede1419bd6f15b9e997ed7a0
Signed-off-by: gbenhaim <gbenhaim@redhat.com>