Add a generic role for deploying operators
When deploying and operator based product, the first step is to ask
OLM to run its operator. This step is the same for all the products,
so it makes sense to have a generic role for it.

The role can be used for different operators by calling it with
different parameters.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>