Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Something is unclear to me about this patch. Is it to accelerate the installation of sriov & performance operators or to fix the issue we have on OCP-4.8 with the MCP of workers ready after the installation?

Hi Charles, sorry for the lack of context here.  While troubleshooting the MCP issues in OCP-4.8 we realized that there were only 2 worker nodes available in cluster5. This setting (maxUnavailable) is set to 2 as a mean to speed up updating the worker nodes. The problem with 2 workers being updated at the same time is that some resources were unable to be evicted as both nodes were unable to take workloads.

While that condition does not fix the issue we were troubleshooting, it solves one scenario (having only 2 workers).  Additionally, when there are more nodes it will set a higher amount of Unavailable nodes to speed up the updates.

Anyway, while going through this explanation I think there's a better approach, I'll submit a patch for that as is IMO better to set the setting to len-1 when >2

Patch-set: 1