Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I don't exactly understand how this patch does what you say it does... is it because you're now prefixing the task so it auto-magically uploads at the end?

In fact, yes ;)
Junit plugin works by consuming env variables that are defined before ansible is launched. That's why those env vars are managed by dci-pipeline. The var JUNIT_TEST_CASE_PREFIX considers a task only as test case if it has this value as prefix and it is set to test_ in dci-pipeline

Patch-set: 1