Include 2M hugepages for load-balancing plugin
This change is needed for enabling the load-balancing plugin. For achieving this:

-We need to set Topology Manager Policy to single-numa-node
-We need 2M hugepages to work properly. The number of pages has also been modified after doing some manual tests, also to meet plugin's workloads. These hugepages have been added in addition to the ones already created for example-cnf, as loadbalancer pod from example-cnf (among others) needs 1G hugepages to work

Moreover, defaultHugepagesSize variable has been removed because it is causing problems with 2M hugepages, which are not installed because defaultSize is 1G.

Finally, some entries from get-logs script has been changed for plugin's jobs. This is a minor change, but included here too.

Change-Id: I12c569c3c91e23503c3ba481b252a99335054516