Standarize tasks to use dci_pullsecret_file var
Currently there are tasks that use dci_pullsecret_file
some others use the path of the clusterX-pull-secret.txt
and some combine registry-auths and pullsecret, so this
change get creds from a single file, create it if it
does not exists, and all tasks that require it use it.
disconnected doc has been updated to reflect these changes.
The content of dci_pullsecret_file is stored in the
pull-secret.txt file in the cache dir for future use.

Change-Id: I83d7e2373c743f22930106142296b8a3cb576629