Some fixes for the sno_on_libvirt
This CR includes some fixes for the SNO solution.

The "create: yes" parameter in task "Add Kubeconfig to Ansible User
.bashrc" in file "sno-installer/tasks/60_deploy_sno_vm.yml"
prevents errors in fresh systems where the .bashrc file may not exist.

We give execution permissions to the oc binary retrieved for the target
version so we can run the binary to extract other needed binaries.

Finally and mainly, we added the "ansible_connection: local" setting to
the provisioner in the SNO sample hosts file so tasks run on the
localhost, like the dci_job tasks, may have access to the local
environment variables, including the DCI credentials.

- Added create clause to line in file in sno_installer
- Fixing execution permissions on oc binary

Test-Hints: sno
Change-Id: Ie4b709f519f0ed7eec447130db55217235065e36