Explicit path and fixes
Overriding the `install_type` in the task makes it so it executes the
task even if it's surrounded by a block with a when condition because
ansible sometimes acts really stupid.
Setting the role path explicitly in ansible.cfg to avoid the issue where
a playbook executed inside the plays/ subdir can't find the other roles
in the ROOT path of d-o-a. Related to https://github.com/dci-labs/dallas-pipelines/pull/424

Change-Id: I2abaae9d34dc625feb3ebc34dae050ce2f5363ea