Reverting to inline teardown
Moving the teardown on error|failure to a play is a good idea in theory,
but the DCI callback is coded to specifically get the include_task
information when presented because of a previous feature request. What
this means is that failures were being shown in the include_task task,
which isn't correct. Having the teardown processes inline in the
playbook and using reusable blocks like before is the best compromise to
not make a lot of repetition.
Read https://issues.redhat.com/browse/CILAB-723 for more context.

Change-Id: I9fd51352b45a187d6f462ce8c552b8b0eaea48ed