Separate performance profile definition from operators installation for CNFs
In OCP 4.11 and 4.12, we are not installing performance-addon-operator but we are installing a performance profile, and it is done because enable_sriov is set to true.

However, this dependency should not be present; we may want to install a performance profile even though we are not using SRIOV.

This change takes care of separating these two elements.

Also, Performance Profile CRD is checked before creating the Performance Profile, so that we do not create it in case the CRD is not present.

Change-Id: I241bb351104db0217ebbc9070b1d4240be474ced