Move config jumpbox to pre-run for all installers
The actions in this configure-jumpbox are required for all installers
directly or indirectly through use of roles.
Moving the task file to the pre-run so is used by all the installers and
removed the tasks covered by it.
Fixing an issue introduced in a previous commit in the AI install
playbook

Change-Id: I1bb46036b602184fea8796667cb0424727655586