Add back a nightly check base page
Despite the efforts to move away from checking the webpage, for
nightlies we still need to do as IPI relies on the release_url. In the
baremetal-deploy the release_url is used in node-prep and node-install
By adding it in here we reduce the failures at the time IPI is being
executed.

Change-Id: I504d259d6d080b1a2826301c5a2723de8d4a2148