Add reliable support to nightly builds
The baremetal-deploy in a connected environment pulls artifacts from a
release_url. This change removes the need to use the url for nightly
builds as is not reliable. Insted, it provides a webserver from the controller
using the already obtained artifacts and provide it to the provisioner
to consume.

Test-Args-Hints: -e dci_topic=OCP-4.12
Change-Id: I2c5ad4fbf5eb16d27ccc92056949c53e159f01ec