filter node names without domain to get logs
this patch removes the node domain if present
and returns the short name, then adds domain
to get logs via ssh/scp. The filter returns
the short node name if no domain is present.
This avoid the domain to be added twice if
the domain is present in the node names from
the inventory

Change-Id: I2f8034531e4ec25b8bb7a4839f0eb7476493e5f0