Fixed failed condition in check-resource/uncordon_workers play
In this job: https://www.distributed-ci.io/jobs/a83deab3-6fb5-4ecc-a17c-70ed73534d42/jobStates?sort=date&task=ffc98df5-a7cf-4645-a380-8040c3843442, this task was failing when checking if the nodes from the cluster are in Ready status and it was failing. By checking the node status, it should not fail because all nodes were in Ready status, in fact.

There was an issue in one of the conditions of this task, where it continues checking the nodes until Ready is not appearing in the output, and it should be NotReady instead

Change-Id: Ib1d6bdaa5fc583f43ac83978639edadcc40bcbec