Increase provisioning timeout
It has been observed in ACM on top on ocp-4.13 a node is declared as provisioned after it the disk image is written and the server is rebooted (around 21mins). In previous version the node was marked as provisioned a couple of minutes after
the BMH is created (2 mins).

So, increasing the timeout for the provisioning completion up to 30 mins.

Change-Id: Ib4b3a237aecf43219914291a9ce9b3116deb36fa