Update usage of SSH keys in testing scripts to follow dci-pipeline approach
The test-runner script in dci-pipeline is not making usage of any SSH key, but ssh-agent is launched in test-runner from dci-openshift-agent.

Updating the code from the testing scripts in this repository to follow the same strategy than in dci-pipeline.

Testing with assisted code to check if SSH key issue is now fixed when retrieving logs from bootstrap machine

Test-Hints: assisted

Change-Id: I419083ef5361563ee11fa34757b1c8c7be984eac