add dns work-around for ocp 4.12 -> 4.13 upgrade
This patch includes tasks to restart pods when the
upgrade is stuck for pods not able to reach the k8s
api default service. Covering the most occurring
scenario with dns pods as described in OCPBUGS-18472
but more scenarios can be added for other pods
because tasks are generic and accept variables.

Test-Upgrade-Hints: yes
Test-Upgrade-From-Topic-Hints: OCP-4.12
Test-Upgrade-To-Topic-Hints: OCP-4.13
Change-Id: I9daef60ce482b3fd2b999747a7ef4efd500ca74f