Confirming Internet connectivity allows 301 status code
The SNO (and the baremetal installer, for that matter), use http queries
to verify the provisioning host has access to the internet.

Usually the http query is sent against www.redhat.com who would respond
with a redirect (301 status code) to www.redhat.com/en.

The URI ansible module used for the check should process the redirection
properly and then get a 200 status code, however in some labs we've seen
that redhat.com keeps responding with 301 even after following the
redirection, which Ansible interprets as an infinity loop, causing the
task to fail.

This change extends the list of accepted status codes to include 301.

Change-Id: Ice04377918ac57ffc0b9b16c063c7b7e28e003b2