unset apply_wa fact in upgrade to avoid WA to run
currently once the fact is set to true, the WA is
executed on every cycle, so we need to update the
value at the end of the WA execution to break the
loop. Also adding a message to print time left to
wait for the upgrade to finish.

Test-Upgrade-Hints: yes
Test-Upgrade-From-Topic-Hints: OCP-4.12
Test-Upgrade-To-Topic-Hints: OCP-4.13
Change-Id: I73a27b617e57525d868c117088a2e44241127f4e