set maxUnavailable worker to n-2 when counter >= 4
This patch modifies default of 1, to speed MC changes
when worker availability is enough to support updates
on 2 nodes at the time. N-2 should work for must use
cases since at least two replicas of any deployment
can be maintained. Also a new variable is defined to
disable it if desired.

Change-Id: I14032acf110217dbb04fecd18cbf420047cba1fb