Increase the disk size for VMs created for Assisted
Test-Hints: assisted-abi

We have seen some warnings during Assisted deployment because of lacking of disk space in the nodes used for deploying OCP. Sometimes, the Assisted deployment fails because of that.

With 250GB disk size, we had no troubles in some tests with partners. Moving to that default disk size

Change-Id: Icad0d56f04dfa5025ecc6e9af8e83271aacb37d6