Retrieving list of cluster nodes by direct method
The get-cluster-facts play identifies the compute and control plane
nodes in the cluster by checking the stored original install-config yaml
file and checking the role assigned to the each node.

This information is later on use, among other things, to add the SRIOV
tag to the compute nodes.

The problem is the stored install-config only represents the initial
configuration of the cluster and, for instance, in SNO deployments, the
install-config does not list a baremetal host at all, so adding the
SRIOV tag would fail.

With this change we get the Node resources from the OCP API and classify
each node based on the labels:

- node-role.kubernetes.io/worker
- node-role.kubernetes.io/master

Change-Id: I71432a7f0c1686358c6eaacb592e638f7942d56f