Add support for performance profile during install
The support of performance profile during install was added in OCP 4.15,
but we are adding it for 4.16+ and moving tuned profile support to its
own play.
We use a predefined name for the performance profile to easily remove in
the cases where is not used in day 1 operations.
Increasing the healthcheck after install as with PAO it will take a bit
longer to apply it to all the nodes.

TestBos2Sno: abi-sno
Change-Id: Ie86fad92925c1b0791a13bef1d7bbc503ceb253d